+++ main.go
@@ -88,7 +88,7 @@ func wrappedMain() int {
 	// Load the configuration
 	config := BuiltinConfig
 	if err := config.Discover(); err != nil {
-		fmt.Fprintf(os.Stderr, "Error discovering plugins: %s", err)
+		Ui.Error(fmt.Sprintf("Error discovering plugins: %s", err))
 		return 1
 	}
 
@@ -123,14 +123,14 @@ func wrappedMain() int {
 	// define extra providers and provisioners.
 	clicfgFile, err := cliConfigFile()
 	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error loading CLI configuration: \n\n%s\n", err)
+		Ui.Error(fmt.Sprintf("Error loading CLI configuration: \n\n%s", err))
 		return 1
 	}
 
 	if clicfgFile != "" {
 		usrcfg, err := LoadConfig(clicfgFile)
 		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error loading CLI configuration: \n\n%s\n", err)
+			Ui.Error(fmt.Sprintf("Error loading CLI configuration: \n\n%s", err))
 			return 1
 		}
 
@@ -143,7 +143,7 @@ func wrappedMain() int {
 
 	exitCode, err := cli.Run()
 	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error executing CLI: %s\n", err.Error())
+		Ui.Error(fmt.Sprintf("Error executing CLI: %s", err.Error()))
 		return 1
 	}
 
@@ -155,7 +155,7 @@ func cliConfigFile() (string, error) {
 	configFilePath := os.Getenv("TERRAFORM_CONFIG")
 	if configFilePath == "" {
 		var err error
-		configFilePath, err = configFile()
+		configFilePath, err = ConfigFile()
 		mustExist = false
 
 		if err != nil {